requirements.txt
fitdecode>=0.11.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
plotly>=5.17.0
jinja2>=3.1.0
python-dateutil>=2.8.0
Pillow>=10.0.0
playwright>=1.40.0

